move methods for flattening netcdf and hdf group structures to separate module
